requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
python-dotenv==1.0.0
openai==1.3.7
langchain==0.0.339
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyMuPDF==1.23.7
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.26.2
google-genai>=1.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
